Enable pins doctests, extend pins API
