+++ Makefile
@@ -77,3 +77,6 @@ clean:
 test:
 	@cargo +nightly test --lib --tests --target $(HOST) --no-default-features --features systick
 	@cargo +nightly test --doc --target $(HOST) --all-features
+
+	@cargo +nightly test --manifest-path teensy4-pins/Cargo.toml --lib --tests --target $(HOST) --all-features
+	@cargo +nightly test --manifest-path teensy4-pins/Cargo.toml --doc --target $(HOST) --all-features
