+++ teensy4-pins/src/t41.rs
@@ -1,6 +1,8 @@
 //! Teensy 4.1 specific APIs
 //!
-//! # Pins unique to the Teensy 4.1
+//! Use [`pins`](fn.pins.html) to constrain the processor pads into the pins available on the Teensy 4.1.
+//! If you cannot safely acquire all processor pads, use the unsafe [`Pins::new`](struct.Pins.html#method.new)
+//! method to generate pins.
 //!
 //! See the [`common` module](../common/index.html) for pins that are consistent
 //! across both boards.
@@ -15,16 +17,6 @@
 //! |  39  |`AD_B1_13`|          |              |              |           |              |                  |              |         |         |         |
 //! |  40  |`AD_B1_04`|          |              |              |           |              |                  |              |         |         |         |
 //! |  41  |`AD_B1_05`|          |              |              |           |              |                  |              |         |         |         |
-//!
-//! # Example
-//!
-//! ```no_run
-//! use teensy4_bsp as bsp;
-//!
-//! let peripherals = bsp::Peripherals::take().unwrap();
-//! let pins = bsp::t41::pins(peripherals.iomuxc);
-//! let led = bsp::configure_led(pins.p13);
-//! ```
 
 pub use crate::common::*;
 use crate::iomuxc::{ad_b1::*, b1::*, ErasedPad};
@@ -147,7 +139,7 @@ pub struct Pins {
 }
 
 /// Constrain the processor pads to the Teensy 4.1 pins
-pub fn pins(iomuxc: crate::iomuxc::Pads) -> Pins {
+pub const fn pins(iomuxc: crate::iomuxc::Pads) -> Pins {
     Pins {
         p0: iomuxc.ad_b0.p03,
         p1: iomuxc.ad_b0.p02,
@@ -196,6 +188,21 @@ pub fn pins(iomuxc: crate::iomuxc::Pads) -> Pins {
 }
 
 impl Pins {
+    /// Create an instance of `Pins` when you do not have a handle
+    /// to the processor pads
+    ///
+    /// # Safety
+    ///
+    /// Caller must ensure that the pins are not aliased elsewhere in
+    /// the program. This could include
+    ///
+    /// - an existing handle to the `imxrt-iomuxc` pads,
+    /// - another instance of `Pins` that was safely acquired
+    ///   using [`pins`](fn.pins.html).
+    pub const unsafe fn new() -> Self {
+        pins(crate::iomuxc::Pads::new())
+    }
+
     /// Erase the types of all pins
     pub fn erase(self) -> ErasedPins {
         [
